Filterbar Hide button
Finalization of Filter Hidden button.

Comments

These are comments to help locate where changes are needed to hide the agent filter.

Set up variables and functions

I created functions and useState variables in the dbUserContext.js file. Also passed the function to the dashboard page to be able to  call on from it.

Function ClassNames changes

There a function running a way to change classNames and working fine. Next is to work on saving to localStorage.

Auto Default Hide Filter

I put classNames to be based on a Ternary condition to hide the filters.
Not to mention defaulted it to hide the side filter.

BugFixing Formcontainer

So I'm going to have to revert the addition to the hiding filter table. This is due to both client and agent having the same container id. So hides both forms.

I'm gonna make a DOM element in javascript and save it to localStorage to save the preference that way it won't affect client side.