BugFixing Tickets
So back then I was asked to only show Open tickets for the front page. I had to make a filter to show the OPEN tickets on the agent dashboard. That was affecting the charts so I had to make another State called allOpenTickets.

This State has all tickets, regardless if resolved.